+++ fyrox-impl/src/engine/mod.rs
@@ -25,11 +25,13 @@
 
 pub mod error;
 pub mod executor;
+pub mod input;
 pub mod task;
 
 mod hotreload;
 mod wasm_utils;
 
+use crate::engine::input::InputState;
 use crate::scene::skybox::SkyBoxKind;
 use crate::{
     asset::{
@@ -130,6 +132,7 @@ use std::{
     time::Duration,
 };
 use uuid::Uuid;
+use winit::event::{DeviceEvent, ElementState, WindowEvent};
 use winit::event_loop::{ActiveEventLoop, EventLoop};
 use winit::{
     dpi::{Position, Size},
@@ -495,6 +498,8 @@ pub struct Engine {
     // Amount of time (in seconds) that passed from creation of the engine.
     elapsed_time: f32,
 
+    input_state: InputState,
+
     /// A special container that is able to create nodes by their type UUID. Use a copy of this
     /// value whenever you need it as a parameter in other parts of the engine.
     pub serialization_context: Arc<SerializationContext>,
@@ -584,6 +589,7 @@ impl ScriptMessageDispatcher {
         user_interfaces: &mut UiContainer,
         graphics_context: &mut GraphicsContext,
         task_pool: &mut TaskPoolHandler,
+        input_state: &InputState,
     ) {
         while let Ok(message) = self.message_receiver.try_recv() {
             let type_id = match message.payload.get_dynamic_type_id() {
@@ -618,6 +624,7 @@ impl ScriptMessageDispatcher {
                                 graphics_context,
                                 user_interfaces,
                                 script_index: 0,
+                                input_state,
                             };
 
                             process_node_scripts(&mut context, &mut |s, ctx| {
@@ -644,6 +651,7 @@ impl ScriptMessageDispatcher {
                                     graphics_context,
                                     user_interfaces,
                                     script_index: 0,
+                                    input_state,
                                 };
 
                                 if receivers.contains(&node) {
@@ -670,6 +678,7 @@ impl ScriptMessageDispatcher {
                                     graphics_context,
                                     user_interfaces,
                                     script_index: 0,
+                                    input_state,
                                 };
 
                                 if receivers.contains(&node) {
@@ -695,6 +704,7 @@ impl ScriptMessageDispatcher {
                                 graphics_context,
                                 user_interfaces,
                                 script_index: 0,
+                                input_state,
                             };
 
                             process_node_scripts(&mut context, &mut |s, ctx| {
@@ -759,6 +769,7 @@ impl ScriptProcessor {
         user_interfaces: &mut UiContainer,
         dt: f32,
         elapsed_time: f32,
+        input_state: &InputState,
     ) {
         self.wait_list
             .retain_mut(|context| !context.is_all_loaded());
@@ -830,6 +841,7 @@ impl ScriptProcessor {
                     graphics_context,
                     user_interfaces,
                     script_index: 0,
+                    input_state,
                 };
 
                 'init_loop: for init_loop_iteration in 0..max_iterations {
@@ -940,6 +952,7 @@ impl ScriptProcessor {
                 user_interfaces,
                 graphics_context,
                 task_pool,
+                input_state,
             );
 
             // As the last step, destroy queued scripts.
@@ -955,6 +968,7 @@ impl ScriptProcessor {
                 graphics_context,
                 task_pool,
                 script_index: 0,
+                input_state,
             };
             while let Some((handle, mut script, index)) = destruction_queue.pop_front() {
                 context.node_handle = handle;
@@ -984,6 +998,7 @@ impl ScriptProcessor {
                     graphics_context,
                     user_interfaces,
                     script_index: 0,
+                    input_state,
                 };
 
                 // Destroy every script instance from nodes that were still alive.
@@ -1240,6 +1255,7 @@ pub(crate) fn process_scripts<T>(
     user_interfaces: &mut UiContainer,
     dt: f32,
     elapsed_time: f32,
+    input_state: &InputState,
     mut func: T,
 ) where
     T: FnMut(&mut Script, &mut ScriptContext),
@@ -1258,6 +1274,7 @@ pub(crate) fn process_scripts<T>(
         graphics_context,
         user_interfaces,
         script_index: 0,
+        input_state,
     };
 
     for node_index in 0..context.scene.graph.capacity() {
@@ -1484,6 +1501,7 @@ impl Engine {
             plugins_enabled: false,
             elapsed_time: 0.0,
             task_pool: TaskPoolHandler::new(task_pool),
+            input_state: Default::default(),
         })
     }
 
@@ -1684,6 +1702,7 @@ impl Engine {
                             async_scene_loader: &mut self.async_scene_loader,
                             loop_controller: controller,
                             task_pool: &mut self.task_pool,
+                            input_state: &self.input_state,
                         };
 
                         for plugin in self.plugins.iter_mut() {
@@ -1717,6 +1736,7 @@ impl Engine {
                     async_scene_loader: &mut self.async_scene_loader,
                     loop_controller: controller,
                     task_pool: &mut self.task_pool,
+                    input_state: &self.input_state,
                 };
 
                 match loading_result.result {
@@ -1913,6 +1933,10 @@ impl Engine {
             self.elapsed_time += dt;
 
             self.post_update_plugins(dt, controller, lag);
+
+            self.input_state.mouse.speed = Vector2::default();
+            self.input_state.keyboard.released_keys.clear();
+            self.input_state.keyboard.pressed_keys.clear();
         }
     }
 
@@ -1939,6 +1963,7 @@ impl Engine {
             &mut self.user_interfaces,
             dt,
             self.elapsed_time,
+            &self.input_state,
         );
 
         self.performance_statistics.scripts_time = instant::Instant::now() - time;
@@ -1971,6 +1996,7 @@ impl Engine {
                         async_scene_loader: &mut self.async_scene_loader,
                         loop_controller: controller,
                         task_pool: &mut self.task_pool,
+                        input_state: &self.input_state,
                     },
                 )
             } else if let Some(node_task_handler) = self.task_pool.pop_node_task_handler(result.id)
@@ -2009,6 +2035,7 @@ impl Engine {
                                         graphics_context: &mut self.graphics_context,
                                         user_interfaces: &mut self.user_interfaces,
                                         script_index: node_task_handler.script_index,
+                                        input_state: &self.input_state,
                                     },
                                 );
 
@@ -2062,6 +2089,7 @@ impl Engine {
                 async_scene_loader: &mut self.async_scene_loader,
                 loop_controller: controller,
                 task_pool: &mut self.task_pool,
+                input_state: &self.input_state,
             };
 
             for plugin in self.plugins.iter_mut() {
@@ -2096,6 +2124,7 @@ impl Engine {
                         async_scene_loader: &mut self.async_scene_loader,
                         loop_controller: controller,
                         task_pool: &mut self.task_pool,
+                        input_state: &self.input_state,
                     };
 
                     for plugin in self.plugins.iter_mut() {
@@ -2132,6 +2161,7 @@ impl Engine {
                 async_scene_loader: &mut self.async_scene_loader,
                 loop_controller: controller,
                 task_pool: &mut self.task_pool,
+                input_state: &self.input_state,
             };
 
             for plugin in self.plugins.iter_mut() {
@@ -2142,13 +2172,69 @@ impl Engine {
         self.performance_statistics.plugins_time += instant::Instant::now() - time;
     }
 
-    pub(crate) fn handle_os_event_by_plugins(
+    /// Should be called on every OS event to update the internal state of the engine accordingly.
+    pub fn handle_os_events(
         &mut self,
         event: &Event<()>,
         dt: f32,
         controller: ApplicationLoopController,
         lag: &mut f32,
     ) {
+        match event {
+            Event::WindowEvent { event, .. } => match event {
+                WindowEvent::KeyboardInput { event, .. } => {
+                    let keyboard = &mut self.input_state.keyboard;
+
+                    match event.state {
+                        ElementState::Pressed => {
+                            if keyboard
+                                .keys
+                                .get(&event.physical_key)
+                                .is_none_or(|state| *state == ElementState::Released)
+                            {
+                                self.input_state
+                                    .keyboard
+                                    .pressed_keys
+                                    .insert(event.physical_key);
+                            }
+                        }
+                        ElementState::Released => {
+                            if keyboard
+                                .keys
+                                .get(&event.physical_key)
+                                .is_some_and(|state| *state == ElementState::Pressed)
+                            {
+                                self.input_state
+                                    .keyboard
+                                    .released_keys
+                                    .insert(event.physical_key);
+                            }
+                        }
+                    }
+
+                    self.input_state
+                        .keyboard
+                        .keys
+                        .insert(event.physical_key, event.state);
+                }
+                WindowEvent::CursorMoved { position, .. } => {
+                    self.input_state.mouse.position =
+                        Vector2::new(position.x as f32, position.y as f32);
+                }
+                _ => (),
+            },
+            Event::DeviceEvent { event, .. } => match event {
+                DeviceEvent::MouseMotion { delta } => {
+                    self.input_state.mouse.speed = Vector2::new(delta.0 as f32, delta.1 as f32);
+                }
+                DeviceEvent::Button { button, state } => {
+                    self.input_state.mouse.buttons_state.insert(*button, *state);
+                }
+                _ => (),
+            },
+            _ => (),
+        }
+
         if self.plugins_enabled {
             for plugin in self.plugins.iter_mut() {
                 plugin.on_os_event(
@@ -2168,6 +2254,7 @@ impl Engine {
                         async_scene_loader: &mut self.async_scene_loader,
                         loop_controller: controller,
                         task_pool: &mut self.task_pool,
+                        input_state: &self.input_state,
                     },
                 );
             }
@@ -2197,6 +2284,7 @@ impl Engine {
                     async_scene_loader: &mut self.async_scene_loader,
                     loop_controller: controller,
                     task_pool: &mut self.task_pool,
+                    input_state: &self.input_state,
                 });
             }
         }
@@ -2225,6 +2313,7 @@ impl Engine {
                     async_scene_loader: &mut self.async_scene_loader,
                     loop_controller: controller,
                     task_pool: &mut self.task_pool,
+                    input_state: &self.input_state,
                 });
             }
         }
@@ -2253,6 +2342,7 @@ impl Engine {
                     async_scene_loader: &mut self.async_scene_loader,
                     loop_controller: controller,
                     task_pool: &mut self.task_pool,
+                    input_state: &self.input_state,
                 });
             }
         }
@@ -2291,6 +2381,7 @@ impl Engine {
                     &mut self.user_interfaces,
                     dt,
                     self.elapsed_time,
+                    &self.input_state,
                     |script, context| {
                         if script.initialized && script.started {
                             script.on_os_event(event, context);
@@ -2382,6 +2473,7 @@ impl Engine {
                             async_scene_loader: &mut self.async_scene_loader,
                             loop_controller: controller,
                             task_pool: &mut self.task_pool,
+                            input_state: &self.input_state,
                         },
                     );
                 }
@@ -2405,6 +2497,7 @@ impl Engine {
                         async_scene_loader: &mut self.async_scene_loader,
                         loop_controller: controller,
                         task_pool: &mut self.task_pool,
+                        input_state: &self.input_state,
                     });
                 }
             }
@@ -2714,6 +2807,7 @@ impl Engine {
             async_scene_loader: &mut self.async_scene_loader,
             loop_controller: controller,
             task_pool: &mut self.task_pool,
+            input_state: &self.input_state,
         });
 
         Log::info(format!("Plugin {plugin_index} was successfully reloaded!"));
@@ -3001,6 +3095,7 @@ mod test {
                 &mut user_interfaces,
                 0.0,
                 0.0,
+                &Default::default(),
             );
 
             match iteration {
@@ -3167,6 +3262,7 @@ mod test {
                 &mut user_interfaces,
                 0.0,
                 0.0,
+                &Default::default(),
             );
 
             match iteration {
@@ -3438,6 +3534,7 @@ mod test {
                 &mut user_interfaces,
                 0.0,
                 0.0,
+                &Default::default(),
             );
 
             match iteration {