+++ fyrox-impl/src/plugin/mod.rs
@@ -24,6 +24,7 @@
 
 pub mod dylib;
 
+use crate::engine::input::InputState;
 use crate::engine::ApplicationLoopController;
 use crate::{
     asset::manager::ResourceManager,
@@ -183,6 +184,14 @@ pub struct PluginContext<'a, 'b> {
 
     /// Task pool for asynchronous task management.
     pub task_pool: &'a mut TaskPoolHandler,
+
+    /// A stored state of most common input events. It is used a "shortcut" in cases where event-based
+    /// approach is too verbose. It may be useful in simple scenarios where you just need to know
+    /// if a button (on keyboard, mouse) was pressed and do something.
+    ///
+    /// **Important:** this structure does not track from which device the corresponding event has
+    /// come from, if you have more than one keyboard and/or mouse, use event-based approach instead!
+    pub input_state: &'a InputState,
 }
 
 define_as_any_trait!(PluginAsAny => Plugin);